feat: 학생회 물건 CRUD 제작

+++ src/main/java/com/example/subsub/domain/CouncilItem.java
@@ -12,7 +12,7 @@ import lombok.*;
 public class CouncilItem {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long itemId;
+    private Integer itemId;
 
     @Column(nullable = false)
     private String name;
